valores receita, despesas e contas (int)

+++ database/migrations/2021_09_02_002057_create_receitas_table.php
@@ -9,7 +9,7 @@ class CreateReceitasTable extends Migration
     public function up(){
         Schema::create('receitas', function (Blueprint $table) {
             $table->increments('id');
-            $table->decimal('valor');
+            $table->integer('valor');
             $table->text('status');
             $table->date('data');
             $table->text('descricao');
